+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 + [x] Tabs: fix clear targets before recollect, issue #1303
 + [x] ListView: fix attributes observing
 + [x] Notify: fix using custom distance option
++ [x] Rating: minor improve code
++ [x] Ribbon menu: fix button group width calc, issue #1296
 
 ### 4.2.34
 + [x] Sidebar: fix error for shifting content issue #1294